Changes to static data

+++ WebApplication1/WebApplication1/Models/Metrics.cs
@@ -5,7 +5,7 @@ namespace CandidView.models
     public class Metrics
     {
         [JsonProperty(PropertyName = "value")]
-        int Value { get; set; }
+        decimal Value { get; set; }
         [JsonProperty(PropertyName = "background")]
         string Background { get; set; }
     }
